// worker/src/types.ts
export interface User {
    id: number;
    user_id: string;
    email: string;
    password?: string;
    quota_size: number;
    used_quota_size: number;
    free_email_time_expire: number;
    role: string;
    created_at: number;
    updated_at: number;
    telegram_chat_id?: string;
    telegram_user_id?: string;
}

export interface EmailRecord {
    id: number;
    user_id: string;
    receiver: string;
    sender: string;
    subject: string;
    body: string;
    created_at: number;
    updated_at: number;
    mail_id: string;
    seen: boolean;
    s3_key?: string;
    sender_ip?: string;
}

// 新增卡密类型
export interface RechargeCode {
    id: number;
    code: string;
    value: number; // 充值天数
    created_at: number;
    used_at: number | null;
    user_id: string | null;
}

// 【新增】域名设置类型
export interface DomainSettings {
    domain: string;
    price: number;
    active: number;
    updated_at: number;
}